new version of philo norminetted

+++ base_functions.c
@@ -6,37 +6,35 @@
 /*   By: aamraouy <aamraouy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/24 12:19:53 by aamraouy          #+#    #+#             */
-/*   Updated: 2025/04/05 17:13:48 by aamraouy         ###   ########.fr       */
+/*   Updated: 2025/04/07 15:03:59 by aamraouy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-		// unsigned int	sign;
-		int	next_digit;
-		int	result;
+	int		next_digit;
+	long	result;
 
-		// sign = 1;
-		result = 0;
-		if(!str)
-			return (0);
-		while (*str == ' ')
-				str++;
-		if ((*str == '-' || *str == '+') && *(str + 1))
-		{
-				if (*str == '-')
-						return (-1);
-				str++;
-		}
-		while (*str >= '0' && *str <= '9')
-		{
-				next_digit = *str - '0';
-				result = result * 10 + next_digit;
-				str++;
-		}
-		return (result);
+	result = 0;
+	if(!str)
+		return (0);
+	while (*str == ' ')
+		str++;
+	if ((*str == '-' || *str == '+') && *(str + 1))
+	{
+		if (*str == '-')
+			return (-1);
+		str++;
+	}
+	while (*str >= '0' && *str <= '9')
+	{
+		next_digit = *str - '0';
+		result = result * 10 + next_digit;
+		str++;
+	}
+	return (result);
 }
 
 int	ft_strlen(char *str)
